feature: added PriceSearchRequest, mapper and started rest controller. refactor: change in PriceId and renamed package mappers
