+++ prices/src/main/java/com/inditex/core/prices/infrastructure/adapters/db/mappers/PriceEntityMapper.java
@@ -1,4 +1,4 @@
-package com.inditex.core.prices.infrastructure.adapters.db.mapper;
+package com.inditex.core.prices.infrastructure.adapters.db.mappers;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
